requirements.txt
fastapi
uvicorn
python-dotenv
streamlit
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
boto3
